yolo sign detection/requirements.txt
ultralytics==8.0.200
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.2
pillow==10.0.0
pyyaml==6.0.1
requests==2.31.0
